temp

+++ Code/Assets/Scripts/Generate/Config/MonsterBase.cs
@@ -86,13 +86,13 @@ namespace Game
 		public string Name { get; set; }
 		/// <summary>攻击</summary>
 		[ProtoMember(4)]
-		public long Attr { get; set; }
+		public double Attr { get; set; }
 		/// <summary>防御</summary>
 		[ProtoMember(5)]
-		public long Def { get; set; }
+		public double Def { get; set; }
 		/// <summary>生命</summary>
 		[ProtoMember(6)]
-		public long HP { get; set; }
+		public double HP { get; set; }
 		/// <summary>DamageIncrea</summary>
 		[ProtoMember(7)]
 		public int DamageIncrea { get; set; }
